New push
I removed some lines in the Method_2 script, as I got feedback on the last assignment that I shouldn't include "install packages" functions in my script.

In the timer script I tried to make a function that spits out what kind of method that is the fastest. This was time consuming, but I think it works in the end. Please see my comments at the end of the script, more specifically the comment on line 77 to 80.